Compact serialization for variable-length integers
Variable-length integers: bytes are a MSB base-128 encoding of the number.
The high bit in each byte signifies whether another digit follows. To make
the encoding is one-to-one, one is subtracted from all but the last digit.
Thus, the byte sequence a[] with length len, where all but the last byte
has bit 128 set, encodes the number:

  (a[len-1] & 0x7F) + sum(i=1..len-1, 128^i*((a[len-i-1] & 0x7F)+1))

Properties:
* Very small (0-127: 1 byte, 128-16511: 2 bytes, 16512-2113663: 3 bytes)
* Every integer has exactly one encoding
* Encoding does not depend on size of original integer type

Former-commit-id: c285438e3724e46c4c0f19f0b72cd180bc7e4d4e [formerly ee6f2a3d3f38b52e055e9ab39b509749d4ffb1db]
Former-commit-id: c285438e3724e46c4c0f19f0b72cd180bc7e4d4e
Former-commit-id: 9abe1991983030a960d0d52247a526afb83b85e8
Former-commit-id: 54f87e72041be9f4e2505c297b6d641d05ff3d9e