RPC, cosmetic: push down ReadHTTPStatus() calls into ReadHTTP() callers
ReadHTTPStatus() is currently overloaded:  In client mode, it properly parses
and receives an HTTP status line.  In server mode, it incorrectly parses the
HTTP request line as an HTTP status line.

This server mode bug has never mattered, because the RPC server never
cared about the URI (path) provided in the HTTP request.  That will change in
the future, so go ahead and begin fixing the problem.

This patch is cosmetic, and should result in NO behavior changes.

Further renames:
ReadHTTPHeader -> ReadHTTPHeaders
ReadHTTP -> ReadHTTPMessage


Former-commit-id: 251a8aaab40b9ad91e13dd2dcb5bfa376e49817c [formerly d4409d191e919e3f63a25c8ec15ddc63bc16a5d5]
Former-commit-id: 251a8aaab40b9ad91e13dd2dcb5bfa376e49817c
Former-commit-id: f0916bb0c2f08ed4342f87434a4b9235b0e3d667
Former-commit-id: b4c537efa1c0a2de7ac76c253fce75ea3c562982