Display tx nLockTime correctly when set to block #
Previously when a transaction was set to lock at a specific block the
calculation was reversed, returning a negative number. This broke the UI
and caused it to display %n in place of the actual number.

In addition the previous calculation would display "Open for 0 blocks"
when the block height was such that the next block created would
finalize the transaction. Inserted the word "more" and changed the
calculation so that the last message would be "Open for 1 more block" to
better match user expectations.

Conflicts:
	src/qt/transactionrecord.h

Former-commit-id: 6adb5304980291841cab27944a201422f0de699a [formerly 5d8aa523d7c69ed50ddfac87d74aef54e67e5379]
Former-commit-id: 6adb5304980291841cab27944a201422f0de699a
Former-commit-id: f6407f9608b915a0cab4d72e686f18836a5676e9
Former-commit-id: 1868288349268196a922ec4382341404512e4738