Treat non-final transactions as non-standard
At least one service that accepted zero-confirmation transactions
was vulnerable because an attacker could send a transaction
with a lock time far in the future, and then have plenty of time in
which to get a double-spend mined (perhaps from a miner who wasn't
on the network when the first transaction was broadcast).

That is a variation on the "Finney attack". We still don't
recommend anybody accept 0-confirmation transactions as final
payment for anything. This change keeps non-final transactions
from appearing in the wallet, and, assuming most of the network
accepts this change, will prevent them from being relayed until
they are final.


Former-commit-id: 614dea07a87703d6ba03825df3c4ac84e5f36139 [formerly 0daa8d9fea07cf38ddaa83062788c2fb67a2b326]
Former-commit-id: 614dea07a87703d6ba03825df3c4ac84e5f36139
Former-commit-id: 43c1a0d25a3ee43a047a9d417583cff98297487a
Former-commit-id: c80259d8c623cf94227cacd717c96036d4e6c800