Make transactions larger than 100K non-standard
Extremely large transactions with lots of inputs can cost the network
almost as much to process as they cost the sender in fees.

We would never create transactions larger than 100K big; this change
makes transactions larger than 100K non-standard, so they are not
relayed/mined by default. This is most important for miners that might
create blocks larger than 250K big, who could be vulnerable to a
make-your-blocks-so-expensive-to-verify-they-get-orphaned attack.

Conflicts:
	src/main.h

Former-commit-id: 0929af06a51cb87fe438b73ff34719c35e3b6d68 [formerly 6485c303c976f41b458050a0a9149a34ad8e7677]
Former-commit-id: 0929af06a51cb87fe438b73ff34719c35e3b6d68
Former-commit-id: 7f4663743083a4dd68ccb3a78c0432a45c126c54
Former-commit-id: 4259d2007df5ec34f9cd79a8c9ed735b72684fba