Internal RNG for approximateBestSubset to prevent degenerate behavior.
This fixes test_bitcoin failures on openbsd reported by dhill on IRC.

  On some systems rand() is a simple LCG over 2^31 and so it produces
an even-odd sequence.  ApproximateBestSubset was only using the least
significant bit and so every run of the iterative solver would be the
same for some inputs, resulting in some pretty dumb decisions.

Using something other than the least significant bit would paper over
the issue but who knows what other way a system's rand() might get us
here.  Instead we use an internal RNG with a period of something like
2^60 which is well behaved.  This also makes it possible to make the
selection deterministic for the tests, if we wanted to implement that.

Conflicts:
	src/test/util_tests.cpp

Former-commit-id: 3a6c256b006e85d12b4cc9ed23854db572913919 [formerly 7c43029d88864774fd18d837d4255c99c3150ef8]
Former-commit-id: 3a6c256b006e85d12b4cc9ed23854db572913919
Former-commit-id: 9bb38e94e39c92f84c8f94df4f15fc1762e0a071
Former-commit-id: fc88bb3d5947a1fdaab92af206dda470699d94c4