Drop release times for CNode
It seems there were two mechanisms for assessing whether a CNode
was still in use: a refcount and a release timestamp. The latter
seems to have been there for a long time, as a safety mechanism.

However, this timer also keeps CNode objects alive for far longer
than necessary after disconnects, potentially opening up a DoS
window.

This commit removes the timestamp-based mechanism, and replaces
it with an assert(nRefCount >= 0), to verify that the refcounting
is indeed correctly working.


Former-commit-id: a9ca85650f59eaa429934dcbba03f80e4badeb9b [formerly 83bd2ec2d377a4aa32ce125e05919d5b62b44579]
Former-commit-id: a9ca85650f59eaa429934dcbba03f80e4badeb9b
Former-commit-id: 364af5f0dd736b07c29b142a5e28b9cc60ef97fc
Former-commit-id: 6306ef99fd8b7270b76b5800f4738449b16e6db5