Reject non-canonically-encoded sizes
The length of vectors, maps, sets, etc are serialized using
Write/ReadCompactSize -- which, unfortunately, do not use a
unique encoding.

So deserializing and then re-serializing a transaction (for example)
can give you different bits than you started with. That doesn't
cause any problems that we are aware of, but it is exactly the type
of subtle mismatch that can lead to exploits.

With this pull, reading a non-canonical CompactSize throws an
exception, which means nodes will ignore 'tx' or 'block' or
other messages that are not properly encoded.

Please check my logic... but this change is safe with respect to
causing a network split. Old clients that receive
non-canonically-encoded transactions or blocks deserialize
them into CTransaction/CBlock structures in memory, and then
re-serialize them before relaying them to peers.

And please check my logic with respect to causing a blockchain
split: there are no CompactSize fields in the block header, so
the block hash is always canonical. The merkle root in the block
header is computed on a vector<CTransaction>, so
any non-canonical encoding of the transactions in 'tx' or 'block'
messages is erased as they are read into memory by old clients,
and does not affect the block hash. And, as noted above, old
clients re-serialize (with canonical encoding) 'tx' and 'block'
messages before relaying to peers.

Conflicts:
	src/main.cpp
	src/test/serialize_tests.cpp

Former-commit-id: b5601cf5bfd080e85ada2bc4e5d75d6151ddfa30 [formerly e097aff4548ee48c8c822015bc81729dd3a799d9]
Former-commit-id: b5601cf5bfd080e85ada2bc4e5d75d6151ddfa30
Former-commit-id: af36df4e32dbf088d26996ee35aa35ffa8cb1683
Former-commit-id: 8ebac4d420f321780cd56250ce741415952a82a9