Wallet locking fixes for -DDEBUG_LOCKORDER
Compiling with -DDEBUG_LOCKORDER and running the qa/rpc-test/ regression
tests uncovered a couple of wallet methods that should (but didn't)
acquire the cs_wallet mutext.

I also changed the AssertLockHeld() routine print to stderr and
abort, instead of printing to debug.log and then assert()'ing.
It is annoying to look in debug.log to find out which
AssertLockHeld is failing.

Conflicts:
	qa/rpc-tests/txnmall.sh
	src/sync.cpp
	src/wallet.cpp

Conflicts:
	src/wallet.cpp

Former-commit-id: 38a1f298411b5adfd01d688d4fc3c8322289c95b [formerly 47083c0435f52623ab8773cad65932b0c52d65e2]
Former-commit-id: 38a1f298411b5adfd01d688d4fc3c8322289c95b
Former-commit-id: df7681856f275f50a7ed51c53a9a92b2c3e27b41
Former-commit-id: 11c760e3ed2c0b32ccda3bf75b12258132f4f88c