Check redeemScript size does not exceed 520 byte limit
redeemScripts >520bytes can't be spent due to the
MAX_SCRIPT_ELEMENT_SIZE limit; previously the addmultisigaddress and
createmultisig RPC calls would let you violate that limit unknowingly.

Also made the wallet code itself check the redeemScript prior to adding
it to the wallet, which in the (rare) instance that a user has added an
invalid oversized redeemScript to their wallet causes an error on
startup. The affected key isn't added to the wallet; other keys are
unaffected.

Conflicts:
	src/keystore.cpp
	src/rpcmisc.cpp
	src/rpcwallet.cpp

Former-commit-id: 56e59cd3a86032a54b7c94d9082128393afcbaa5 [formerly 90f5f659e32cee1e8f6f68616f1a21d244eb7e26]
Former-commit-id: 56e59cd3a86032a54b7c94d9082128393afcbaa5
Former-commit-id: bdc7cbed37f4477c22ff2bf9ed86ed2953b57047
Former-commit-id: 080b1807f1785c2c8760a351edd0f1e55f1cc884