Remove NumBlocksOfPeers
Generally useless information. Only updates on connect time, not after
that. Peers can easily lie and the median filter is not effective in
preventing that.

In the past it was used for progress display in the GUI but
`CheckPoints::guessVerificationProgress` provides a better way that is now used.
It was too easy to mislead it. Peers do lie about it in practice, see issue #4065.

From the RPC, `getpeerinfo` gives the peer raw values, which are more
useful.

Conflicts:
	src/main.cpp
	src/main.h
	src/qt/bitcoingui.h
	src/qt/clientmodel.cpp
	src/qt/clientmodel.h
	src/qt/forms/rpcconsole.ui
	src/qt/rpcconsole.cpp

Conflicts:
	src/qt/clientmodel.h

Former-commit-id: 7d9eee97efd81f6efe297b17a466c69f637ba230 [formerly b921090b08df55339b8e30f9db5d08d5be95369a]
Former-commit-id: 7d9eee97efd81f6efe297b17a466c69f637ba230
Former-commit-id: 96aaa1d53f7f54bdb3a1ef87c36148a2a6bcef17
Former-commit-id: 0bb3ffde8a2c6ac99bc7daf7d2f89d6cc4170b3d