key.cpp: fail with a friendlier message on missing ssl EC support
Previously if bitcoind is linked with an OpenSSL which is compiled
without EC support, this is seen as an assertion failure "pKey !=
NULL" at key.cpp:134, which occurs after several seconds. It is an
esoteric piece of knowledge to interpret this as "oops, I linked
with the wrong OpenSSL", and because of the delay it may not even
be noticed.

The new output is

: OpenSSL appears to lack support for elliptic curve cryptography. For
more information, visit
https://en.bitcoin.it/wiki/OpenSSL_and_EC_Libraries
: Initialization sanity check failed. Bitcoin Core is shutting down.

which occurs immediately after attempted startup.

This also blocks in an InitSanityCheck() function which currently only
checks for EC support but should eventually do more. See #4081.

Conflicts:
	src/init.cpp
	src/key.h

Former-commit-id: 448d6d9aecf1a28b72b1803f81b9fd25cb8630fd [formerly a5ce629b62e3ab732ad05b649ba1125a9cc5c696]
Former-commit-id: 448d6d9aecf1a28b72b1803f81b9fd25cb8630fd
Former-commit-id: 26f336f49830b391827b4f1f15201ffdd81c91bb
Former-commit-id: d31651f92574addffccae2ca0be06e938e80a1bc